Tests for some instruments added.
